extract chat template detection into SRP microcrate
