+++ crates/bitnet-chat-template-detection-core/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "bitnet-chat-template-detection-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP chat template auto-detection primitives"
+
+[dependencies]
+
+[lints]
+workspace = true
