+++ crates/bitnet-cli/src/commands/mod.rs
@@ -9,7 +9,6 @@ pub mod eval;
 pub mod inference;
 pub mod inspect;
 pub mod serve;
-pub mod template_util;
 
 #[cfg(feature = "cli-bench")]
 pub use benchmark::BenchmarkCommand;
